support for vitis 2023.1

+++ networklayer/synthesis_results_HBM/ARP_hls_prj/ultrascale_plus/impl/misc/drivers/arp_server_v1_0/src/xarp_server.h
@@ -1,6 +1,9 @@
 // ==============================================================
-// Vitis HLS - High-Level Synthesis from C, C++ and OpenCL v2021.2 (64-bit)
-// Copyright 1986-2021 Xilinx, Inc. All Rights Reserved.
+// Vitis HLS - High-Level Synthesis from C, C++ and OpenCL v2023.1 (64-bit)
+// Tool Version Limit: 2023.05
+// Copyright 1986-2022 Xilinx, Inc. All Rights Reserved.
+// Copyright 2022-2023 Advanced Micro Devices, Inc. All Rights Reserved.
+// 
 // ==============================================================
 #ifndef XARP_SERVER_H
 #define XARP_SERVER_H
@@ -38,7 +41,7 @@ typedef uint64_t u64;
 #else
 typedef struct {
     u16 DeviceId;
-    u32 S_axilite_BaseAddress;
+    u64 S_axilite_BaseAddress;
 } XArp_server_Config;
 #endif
