+++ networklayer/synthesis_results_HBM/UDP_prj/ultrascale_plus/impl/misc/drivers/udp_v1_0/src/xudp.h
@@ -1,6 +1,9 @@
 // ==============================================================
-// Vitis HLS - High-Level Synthesis from C, C++ and OpenCL v2021.2 (64-bit)
-// Copyright 1986-2021 Xilinx, Inc. All Rights Reserved.
+// Vitis HLS - High-Level Synthesis from C, C++ and OpenCL v2023.1 (64-bit)
+// Tool Version Limit: 2023.05
+// Copyright 1986-2022 Xilinx, Inc. All Rights Reserved.
+// Copyright 2022-2023 Advanced Micro Devices, Inc. All Rights Reserved.
+// 
 // ==============================================================
 #ifndef XUDP_H
 #define XUDP_H
@@ -38,7 +41,7 @@ typedef uint64_t u64;
 #else
 typedef struct {
     u16 DeviceId;
-    u32 S_axilite_BaseAddress;
+    u64 S_axilite_BaseAddress;
 } XUdp_Config;
 #endif
